start changing the whole code structure
