+++ src/sgchemist/orm/typing_util.py
@@ -1,5 +1,5 @@
 """Collections of typing utility functions."""
-
+from __future__ import absolute_import
 from __future__ import annotations
 
 import builtins
