+++ tests/test_orm/classes.py
@@ -6,15 +6,15 @@ from typing import List
 from typing import Optional
 from typing import Union
 
+from sgchemist.orm.descriptor import alias_relationship
+from sgchemist.orm.descriptor import field
+from sgchemist.orm.descriptor import relationship
 from sgchemist.orm.entity import SgEntity
-from sgchemist.orm.field import DateTimeField
-from sgchemist.orm.field import EntityField
-from sgchemist.orm.field import ImageField
-from sgchemist.orm.field import MultiEntityField
-from sgchemist.orm.field import TextField
-from sgchemist.orm.field_descriptor import alias_relationship
-from sgchemist.orm.field_descriptor import mapped_field
-from sgchemist.orm.field_descriptor import relationship
+from sgchemist.orm.fields import DateTimeField
+from sgchemist.orm.fields import EntityField
+from sgchemist.orm.fields import ImageField
+from sgchemist.orm.fields import MultiEntityField
+from sgchemist.orm.fields import TextField
 
 
 class Project(SgEntity):
@@ -28,7 +28,7 @@ class Shot(SgEntity):
     """A test shot entity."""
 
     __sg_type__ = "Shot"
-    name: TextField = mapped_field("code", name_in_relation="name")
+    name: TextField = field("code", name_in_relation="name")
     description: TextField
     project: EntityField[Project]
     parent_shots: MultiEntityField[List[Shot]]
@@ -40,7 +40,7 @@ class Asset(SgEntity):
     """A test asset entity."""
 
     __sg_type__ = "Asset"
-    name: TextField = mapped_field("code", name_in_relation="name")
+    name: TextField = field("code", name_in_relation="name")
     project: EntityField[Project]
     shots: MultiEntityField[List[Shot]]
     tasks: MultiEntityField[List[Task]]
@@ -50,9 +50,9 @@ class Task(SgEntity):
     """A test task entity."""
 
     __sg_type__ = "Task"
-    name: TextField = mapped_field(name="content")
+    name: TextField = field(name="content")
     entity: EntityField[Optional[Union[Asset, Shot]]] = relationship()
     shot: EntityField[Optional[Shot]] = alias_relationship(entity)
     asset: EntityField[Optional[Asset]] = alias_relationship(entity)
     created_at: DateTimeField
-    image: ImageField = mapped_field()
+    image: ImageField = field()
