+++ tests/test_orm/test_operator.py
@@ -6,15 +6,15 @@ import pytest
 from sgchemist.orm.constant import LogicalOperator
 from sgchemist.orm.constant import Operator
 from sgchemist.orm.entity import SgEntity
+from sgchemist.orm.field import AbstractValueField
 from sgchemist.orm.field import TextField
-from sgchemist.orm.instrumentation import InstrumentedField
 from sgchemist.orm.queryop import SgFieldCondition
 from sgchemist.orm.queryop import SgFilterOperation
 from sgchemist.orm.queryop import SgNullCondition
 
 
 @pytest.fixture
-def field() -> InstrumentedField[Any]:
+def field() -> AbstractValueField[Any]:
     """Returns the test field."""
 
     class _TestModel(SgEntity):
@@ -24,7 +24,7 @@ def field() -> InstrumentedField[Any]:
     return _TestModel.test
 
 
-def test_field_condition(field: InstrumentedField[Any]) -> None:
+def test_field_condition(field: AbstractValueField[Any]) -> None:
     """Tests field condition."""
     cond1 = SgFieldCondition(field, Operator.IS, "foo")
     cond2 = SgFieldCondition(field, Operator.IS, "name")
@@ -38,7 +38,7 @@ def test_field_condition(field: InstrumentedField[Any]) -> None:
     assert or_cond.sg_objects == [cond1, cond2]
 
 
-def test_filter_operator(field: InstrumentedField[Any]) -> None:
+def test_filter_operator(field: AbstractValueField[Any]) -> None:
     """Tests filter operator."""
     cond1 = SgFieldCondition(field, Operator.IS, "foo")
     cond2 = SgFieldCondition(field, Operator.IS, "name")
@@ -72,7 +72,7 @@ def test_filter_operator(field: InstrumentedField[Any]) -> None:
     assert or_concat_or.sg_objects == [cond1, cond2, cond1, cond2]
 
 
-def test_null_condition(field: InstrumentedField[Any]) -> None:
+def test_null_condition(field: AbstractValueField[Any]) -> None:
     """Tests null condition."""
     null_cond = SgNullCondition()
     self_null = null_cond & SgNullCondition()
